fix(Button): better type of props (additional to e7cd47e68327dea698aa8a83ed06a3b5cef13882)
Unfortunately typescript cannot determine which type is used when we
don't provide `href` property. As result it brings some compilation
errors with `ref` and `component` fields.

So we need to bring back the conditional inferring of props type.
It was a mistake to think we can avoid it.

Additionaly `Exclude<ElementType, 'a' | 'button'>` is extracted
as separate type `ButtonCustomElementType`